# api/api.yml
...
openapi: 3.1.0
info:
  version: 1.0.0
  title: Typhoon API
  summary: |-
    Typhoon is an event bridge built on top of an event mesh,
    with a powerful API and beautiful experience.
  contact:
    name: Sebastian Doell
    email: sebastian@katallaxie.dev
  description: |
    # Introduction

    Typhoon is an event bridge for applications and services.
  license:
    name: Apache 2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
paths:
...
  /operators:
    get:
      summary: List all operators
      operationId: listOperators
      tags:
        - Operator
      parameters:
        - $ref: "#/components/parameters/offsetParam"
        - $ref: "#/components/parameters/limitParam"
      security:
        - bearerAuth: ["read:operators"]
        - apiKey: ["superadmin"]
      responses:
        200:
          description: Successfull response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResult"
                  - $ref: "#/components/schemas/Operators"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
    post:
      summary: Creates a new operator
      operationId: createOperator
      tags:
        - Operator
      requestBody:
        $ref: "#/components/requestBodies/CreateOperator"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
  /operators/{operatorId}:
    get:
      summary: Gets an operator by ID
      operationId: getOperator
      parameters:
        - $ref: "#/components/parameters/operatorId"
      responses:
        200:
          description: Successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
    put:
      summary: Updates an operator by ID
      operationId: updateOperator
      parameters:
        - $ref: "#/components/parameters/operatorId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Operator"
      responses:
        200:
          description: Successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
    delete:
      summary: Deletes an operator by ID
      operationId: deleteOperator
      parameters:
        - $ref: "#/components/parameters/operatorId"
      responses:
        204:
          description: Successful
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
  /operators/{operatorId}/groups:
    get:
      summary: List all signing key groups for an operator
      operationId: listOperatorSigningKeyGroups
      tags:
        - Operator
      parameters:
        - $ref: "#/components/parameters/operatorId"
        - $ref: "#/components/parameters/offsetParam"
        - $ref: "#/components/parameters/limitParam"
      security:
        - bearerAuth: []
        - apiKey: []
      responses:
        200:
          description: Successfull response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResult"
                  - type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/SigningKeyGroup"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
    post:
      summary: Creates a new signing key group
      operationId: createOperatorSigningKeyGroup
      tags:
        - Operator
      parameters:
        - $ref: "#/components/parameters/operatorId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SigningKeyGroup"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SigningKeyGroup"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
...
  /operators/{operatorId}/accounts:
    get:
      summary: List all accounts for an operator
      operationId: listOperatorAccounts
      tags:
        - Operator
      parameters:
        - $ref: "#/components/parameters/operatorId"
        - $ref: "#/components/parameters/offsetParam"
        - $ref: "#/components/parameters/limitParam"
      security:
        - bearerAuth: ["read:accounts"]
      responses:
        200:
          description: Successfull response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResult"
                  - type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/Account"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
    post:
      summary: Creates a new account
      operationId: createOperatorAccount
      tags:
        - Operator
      parameters:
        - $ref: "#/components/parameters/operatorId"
      requestBody:
        $ref: "#/components/requestBodies/CreateAccount"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
  /operators/{operatorId}/accounts/{accountId}:
    get:
      summary: Gets an account by ID
      operationId: getOperatorAccount
      parameters:
        - $ref: "#/components/parameters/operatorId"
        - $ref: "#/components/parameters/accountId"
      responses:
        200:
          description: Successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"

    put:
      summary: Updates an account by ID
      operationId: updateOperatorAccount
      parameters:
        - $ref: "#/components/parameters/operatorId"
        - $ref: "#/components/parameters/accountId"
      requestBody:
        $ref: "#/components/requestBodies/UpdateAccount"
      responses:
        200:
          description: Successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"

    delete:
      summary: Deletes an account by ID
      operationId: deleteOperatorAccount
      parameters:
        - $ref: "#/components/parameters/operatorId"
        - $ref: "#/components/parameters/accountId"
      responses:
        204:
          description: Successful
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
  /operators/{operatorId}/accounts/{accountId}/token:
    get:
      summary: Gets a token for an account
      operationId: getOperatorAccountToken
      parameters:
        - $ref: "#/components/parameters/operatorId"
        - $ref: "#/components/parameters/accountId"
      responses:
        200:
          description: Successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JWTToken"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
    delete:
      summary: Deletes a token for an account
      operationId: deleteOperatorAccountToken
      parameters:
        - $ref: "#/components/parameters/operatorId"
        - $ref: "#/components/parameters/accountId"
      responses:
        204:
          description: Successful
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
  /operators/{operatorId}/accounts/{accountId}/users:
    get:
      summary: List all users for an account
      operationId: listOperatorAccountUsers
      tags:
        - Operator
      parameters:
        - $ref: "#/components/parameters/operatorId"
        - $ref: "#/components/parameters/accountId"
        - $ref: "#/components/parameters/offsetParam"
        - $ref: "#/components/parameters/limitParam"
      security:
        - bearerAuth: ["read:users"]
      responses:
        200:
          description: Successfull response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResult"
                  - type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/User"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
    post:
      summary: Creates a new user
      operationId: createOperatorAccountUser
      tags:
        - Operator
      parameters:
        - $ref: "#/components/parameters/operatorId"
        - $ref: "#/components/parameters/accountId"
      requestBody:
        $ref: "#/components/requestBodies/CreateOperatorAccountUser"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
...
  /operators/{operatorId}/accounts/{accountId}/users/{userId}/credentials:
    get:
      summary: Gets the credentials for a user
      operationId: getOperatorAccountUserCredentials
      parameters:
        - $ref: "#/components/parameters/operatorId"
        - $ref: "#/components/parameters/accountId"
        - $ref: "#/components/parameters/userId"
      responses:
        200:
          description: Successful
          content:
            application/octet-stream:
              schema:
                $ref: "#/components/schemas/Credentials"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
...
  /teams/{teamId}/accounts/{accountId}/users:
    get:
      summary: List all users for an account
      operationId: listUsers
      tags:
        - Team
      parameters:
        - $ref: "#/components/parameters/teamId"
        - $ref: "#/components/parameters/accountId"
        - $ref: "#/components/parameters/offsetParam"
        - $ref: "#/components/parameters/limitParam"
      security:
        - bearerAuth: ["read:users"]
      responses:
        200:
          description: Successfull response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResult"
                  - type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/User"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
    post:
      summary: Creates a new user
      operationId: createUser
      tags:
        - Team
      parameters:
        - $ref: "#/components/parameters/teamId"
        - $ref: "#/components/parameters/accountId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
...
  /systems/{systemId}:
    get:
      summary: Gets a system by ID
      operationId: getSystem
      parameters:
        - $ref: "#/components/parameters/systemId"
      responses:
        200:
          description: Successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/System"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
    put:
      summary: Updates a system by ID
      operationId: updateSystem
      parameters:
        - $ref: "#/components/parameters/systemId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/System"
      responses:
        200:
          description: Successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/System"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"

    delete:
      summary: Deletes a system by ID
      operationId: deleteSystem
      parameters:
        - $ref: "#/components/parameters/systemId"
      responses:
        204:
          description: Successful
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
  /systems/{systemId}/operator:
    get:
      summary: Gets the operator for a system
      operationId: getSystemOperator
      parameters:
        - $ref: "#/components/parameters/systemId"
      responses:
        200:
          description: Successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
    put:
      summary: Updates the operator for a system
      operationId: updateSystemOperator
      parameters:
        - $ref: "#/components/parameters/systemId"
      requestBody:
        $ref: "#/components/requestBodies/UpdateSystemOperator"
      responses:
        201:
          description: Successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
    delete:
      summary: Deletes the operator for a system
      operationId: deleteSystemOperator
      parameters:
        - $ref: "#/components/parameters/systemId"
      responses:
        204:
          description: Successful
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "./schemas/error.yml#/components/schemas/Error"
components:
  parameters:
...
    accountId:
      in: path
      name: accountId
      required: true
      schema:
        type: string
        format: uuid
      description: The id of the account to retrieve.
...
    offsetParam:
      in: query
      name: offset
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: The number of items to skip before starting to collect the result set.
    limitParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
      description: The numbers of items to return.
    operatorId:
      in: path
      name: operatorId
      required: true
      schema:
        type: string
        format: uuid
      description: The id of the operator to retrieve.
...
  requestBodies:
...
    CreateOperator:
      content:
        application/json:
          schema:
            type: object
            required:
              - name
            properties:
              name:
                type: string
              description:
                type: string
              contactEmail:
                type: string
    # Body for creating an account
    CreateAccount:
      content:
        application/json:
          schema:
            type: object
            required:
              - name
            properties:
              name:
                type: string
              description:
                type: string

    # Body for updating an account
    UpdateAccount:
      content:
        application/json:
          schema:
            type: object
            required:
              - name
            properties:
              name:
                type: string
              description:
                type: string
              claims:
                $ref: "#/components/schemas/JWTAccountClaims"
    # Body for creating an operator account user
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    PaginatedResult:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        results:
          type: array
          items:
...
    Operator:
      type: object
      required:
        - name
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        key:
          $ref: "#/components/schemas/KeyPair"
        signingKeys:
          type: array
          items:
            $ref: "#/components/schemas/KeyPair"
        description:
          type: string
        contactEmail:
          type: string
        createdAt:
          type: string
          format: date-time
          description: Creation date and time
          example: "2021-01-30T08:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Creation date and time
          example: "2021-01-30T08:30:00Z"
        deletedAt:
          type: string
          format: date-time
          description: Creation date and time
          example: "2021-01-30T08:30:00Z"
...
    Account:
      type: object
      required:
        - name
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        description:
          type: string
        contactEmail:
          type: string
        key:
          $ref: "#/components/schemas/KeyPair"
        signingKeys:
          type: array
          items:
            $ref: "#/components/schemas/KeyPair"
        createdAt:
          type: string
          format: date-time
          description: Creation date and time
          example: "2021-01-30T08:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Creation date and time
          example: "2021-01-30T08:30:00Z"
        deletedAt:
          type: string
          format: date-time
          description: Creation date and time
          example: "2021-01-30T08:30:00Z"
    SigningKeyGroup:
      type: object
      required:
        - name
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        description:
          type: string
        scope:
          $ref: "#/components/schemas/SigningKeyGroupScope"
        is_scoped:
          type: boolean
          default: false
        disabled:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
          description: Creation date and time
          example: "2021-01-30T08:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Creation date and time
          example: "2021-01-30T08:30:00Z"
        deletedAt:
          type: string
          format: date-time
          description: Creation date and time
          example: "2021-01-30T08:30:00Z"
...
    Operators:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/Operator"
...
security:
  - bearerAuth: []
  - apiKey: []